Add User entity, write repository (spring Data) for User, add Integra…

+++ cafiteria_back/src/main/resources/db/migration/V00001__add_user_table.sql
@@ -8,7 +8,7 @@ CREATE TABLE users
     name VARCHAR(255) NOT NULL,
     surname VARCHAR(255) NOT NULL,
     location VARCHAR(255) NOT NULL,
-    phone INT NOT NULL UNIQUE ,
+    phone BIGINT NOT NULL UNIQUE ,
     email VARCHAR(255) NOT NULL UNIQUE ,
     password VARCHAR(255) NOT NULL,
     PRIMARY KEY (id)
